tsv git clean option

+++ eng/scripts/Validate-TypeSpec.ps1
@@ -5,7 +5,9 @@ param (
   [Parameter(Position = 1, Mandatory = $false)]
   [string]$TargetBranch,
   [Parameter(Position = 2, Mandatory = $false)]
-  [string]$SourceBranch
+  [string]$SourceBranch,
+  [Parameter(Position = 2, Mandatory = $false)]
+  [switch]$GitClean
 )
 
 $exitCode = 0
@@ -25,8 +27,10 @@ if ($typespecFolders) {
     if ($LASTEXITCODE) {
       $exitCode = 1
     }
-    git restore .
-    git clean -df
+    if ($GitClean) {
+      git restore .
+      git clean -df
+    }
   }
 }
